+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version 2.5.22
 ## Major Changes
 * Updated AIC GoF to run lognormal and voom for nondiscrete data
+* Added AIC_median value
+* Updated compute_aic to return a dataframe (rather than multiple) and a boxplox
 
 # Version 2.5.21
 ## Major Changes